+++ posthog/src/main/java/com/posthog/internal/PostHogFeatureFlags.kt
@@ -2,6 +2,8 @@ package com.posthog.internal
 
 import com.posthog.PostHogConfig
 import com.posthog.PostHogOnFeatureFlags
+import com.posthog.internal.PostHogPreferences.Companion.FEATURE_FLAGS
+import com.posthog.internal.PostHogPreferences.Companion.FEATURE_FLAGS_PAYLOAD
 import java.util.concurrent.ExecutorService
 import java.util.concurrent.atomic.AtomicBoolean
 
@@ -63,13 +65,17 @@ internal class PostHogFeatureFlags(
                             val normalizedPayloads = normalizePayloads(response.featureFlagPayloads)
                             this.featureFlagPayloads = normalizedPayloads
                         }
+                        config.cachePreferences?.let { preferences ->
+                            val flags = this.featureFlags ?: mapOf()
+                            preferences.setValue(FEATURE_FLAGS, flags)
+
+                            val payloads = this.featureFlagPayloads ?: mapOf()
+                            preferences.setValue(FEATURE_FLAGS_PAYLOAD, payloads)
+                        }
                     }
                     isFeatureFlagsLoaded = true
-                } ?: run {
-                    isFeatureFlagsLoaded = false
                 }
             } catch (e: Throwable) {
-                isFeatureFlagsLoaded = false
                 config.logger.log("Loading feature flags failed: $e")
             } finally {
                 try {
@@ -78,6 +84,31 @@ internal class PostHogFeatureFlags(
                     config.logger.log("Executing the feature flags callback failed: $e")
                 } finally {
                     isLoadingFeatureFlags.set(false)
+                }
+            }
+        }
+    }
+
+    private fun loadFeatureFlagsFromCache() {
+        synchronized(featureFlagsLock) {
+            config.cachePreferences?.let { preferences ->
+                synchronized(featureFlagsLock) {
+                    @Suppress("UNCHECKED_CAST")
+                    val flags = preferences.getValue(
+                        FEATURE_FLAGS,
+                        mapOf<String, Any>(),
+                    ) as? Map<String, Any> ?: mapOf()
+
+                    @Suppress("UNCHECKED_CAST")
+                    val payloads = preferences.getValue(
+                        FEATURE_FLAGS_PAYLOAD,
+                        mapOf<String, Any?>(),
+                    ) as? Map<String, Any?> ?: mapOf()
+
+                    this.featureFlags = flags
+                    this.featureFlagPayloads = payloads
+
+                    isFeatureFlagsLoaded = true
                 }
             }
         }
@@ -106,7 +137,7 @@ internal class PostHogFeatureFlags(
 
     fun isFeatureEnabled(key: String, defaultValue: Boolean): Boolean {
         if (!isFeatureFlagsLoaded) {
-            return defaultValue
+            loadFeatureFlagsFromCache()
         }
         val value: Any?
 
@@ -128,7 +159,7 @@ internal class PostHogFeatureFlags(
 
     private fun readFeatureFlag(key: String, defaultValue: Any?, flags: Map<String, Any?>?): Any? {
         if (!isFeatureFlagsLoaded) {
-            return defaultValue
+            loadFeatureFlagsFromCache()
         }
         val value: Any?
 
@@ -159,6 +190,11 @@ internal class PostHogFeatureFlags(
         synchronized(featureFlagsLock) {
             featureFlags = null
             featureFlagPayloads = null
+
+            config.cachePreferences?.let { preferences ->
+                preferences.remove(FEATURE_FLAGS)
+                preferences.remove(FEATURE_FLAGS_PAYLOAD)
+            }
         }
     }
 }